Fixed issue with listener
When I tried the page on the browser, it would print all the information default. I believe it was because I was calling the function like this: showDetails(pokemon) so it would trigger it as I loaded the page... so changed it to have the function inside the event function.